some contributions endpoints

+++ src/main/java/com/erickmob/yieldme/repository/WalletRepository.java
@@ -4,7 +4,7 @@ import com.erickmob.yieldme.model.User;
 import com.erickmob.yieldme.model.Wallet;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WalletRepository extends JpaRepository<Wallet, Integer> {
+public interface WalletRepository extends JpaRepository<Wallet, Long> {
 
     Wallet findByUser(User user);
